0.0.3 release version add gif and net downloader

+++ MFMediaView/Classes/MFMediaView/MFMediaViewModelPAGConfig.m
@@ -29,7 +29,7 @@
 - (NSString *)description {
     NSMutableString *description = [NSMutableString stringWithFormat:@"<%@: ", NSStringFromClass([self class])];
     [description appendFormat:@"self.repeatCount=%lu", self.repeatCount];
-    [description appendFormat:@", self.scaleMode=%d", self.scaleMode];
+    [description appendFormat:@", self.scaleMode=%ld", (long)self.scaleMode];
     [description appendFormat:@", self.maxFrameRate=%lu", self.maxFrameRate];
     [description appendFormat:@", self.onAnimateStopAction=%p", self.onAnimateStopAction];
     [description appendString:@">"];
